requirements.txt
streamlit==1.32.0
openai==1.12.0
python-dotenv==1.0.1
python-docx==1.1.0
langchain-core>=0.3
langchain-community>=0.3
langchain-text-splitters>=0.2
faiss-cpu==1.7.4
tiktoken>=0.5.1
plotly==5.18.0
pdfplumber==0.10.3
matplotlib 